requirements.txt
jupyter
ipykernel
notebook
colorama
ipdb
matplotlib
seaborn
numpy
pandas
scipy
# To balance classes with SMOTE
imblearn
scikit-learn
pytest
# GCP
google-cloud-storage
google-cloud-bigquery
db-dtypes # For BigQuery to convert pandas dtypes to BigQuery dtypes
# API
fastapi
uvicorn
pydantic
# dotenv
python-dotenv